tdf#132557: PPTX import: Workaround for slide footer shape presets
It appears that placeholder shapes with service names:
  - com.sun.star.presentation.TitleTextShape
  - com.sun.star.presentation.DateTimeShape
  - com.sun.star.presentation.FooterShape
  - com.sun.star.presentation.SlideNumberShape
  ...

Are unable to have custom shapes in Impress. (i.e. can't be
ellipse, triangle etc.). These presets get specified in OOXML
with <a:prstGeom prst="ellipse"/> inside spPr (shape properties).

Therefore with similar results to the PPT import, a workaround
is applied where slide footers which have non default shapes
are imported with com.sun.star.drawing.CustomShapes service.

The layout/master footers are left as is since if they were to
be imported as CustomShapes they would appear on each slide
even if the slide had those footers enabled or not.

Change-Id: Ic8a8ab3f6dfb7367ecd2c619ce888bf77abef460
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129889
Tested-by: Jenkins
Reviewed-by: Jan Holesovsky <kendy@collabora.com>