tdf#95882 svxruler: restore adjustment for borders
This restores the clause removed from
LO 3.6's commit 0d635c8e9440fafeb63cee1d6d201334439eadfe
    Ruler: disable snapping, tooltips

[Because that removal left behind dead code,
 which was later cleared out, of course this patch
 restores that as well.
    commit aeff565a0be6b810eb4412e71b50ed464ce2fd42
    Author: Noel Grandin on Fri Jun 14 10:41:11 2019 +0200
]

Basic borders on the paragraph itself
do not affect where LO ends the tabstop,
and thus the removal of that clause was partially justified.
But the table/frame/page/headers definitely need to be adjusted
in order for the tabstops to show in the correct position.

RightFrameMargin needed this as well,
especially now that RTL is supported by the ruler since LO 3.6,
although the Left/Right functions are very different.

[Frames never worked because the aCoreSet wasn't accepting
 RES_BOX items. Interesting error considering all of the
 other sections of code created a new CoreSet.]

Change-Id: Ic24839dbbd730e66d8b0d588bfbd73bc7cb260bf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130125
Tested-by: Justin Luth <jluth@mail.com>
Reviewed-by: Justin Luth <jluth@mail.com>