uilogger: make vertical tab handling more robust
For example hyperlink dialog does not have numeric ids for tabs.
Now the rule matches both integers and "common identifiers" as per
https://textx.github.io/textX/stable/grammar/#textx-base-types

Parent argument was removed as it does not seem to be used and
it was empty in the case of the hyperlink dialog, causing the DSL
parser to fail.

More work is needed as self.xUITest.getTopFocusWindow() in a
hyperlink dialog opening test does not contain the child "tabcontrol"
and thus it errors out.

Change-Id: I1637f7300884881ee4d869e8f04700e56935f2d7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130111
Tested-by: Jenkins
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>