tdf#147166 EditTextObjectImpl copy ctor doesn't exactly copy EditTextObjectImpl
and this is apparently relied on, so eliding the copy gives unexpected
results.

EditTextObjectImpl::Clone returns a copy of *this, but the
EditTextObjectImpl copy ctor explicitly does not copy the "PortionInfo"
member, so in:

commit 963a8599d1dbab1b33ca106598d87ab28678729b
Author: Caolán McNamara <caolanm@redhat.com>
Date:   Sat Oct 10 19:21:38 2020 +0100

    ofz#23492 the only user of this ctor throws away the original of the clone

    so we can take ownership of the original instead

where the copy was optimized away we want from a state where there was a
new EditTextObjectImpl with an empty PortionInfo member to one where the
PortionInfo of the EditTextObjectImpl was retained.

So explicitly clear this unwanted info.

It's very hard to make rational judgements about code if a copy behaves
differently than the orignal :-(

Change-Id: I642d60841d6bdccbf830f8a2ccdbd9f542a8aa18
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130201
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>