make sure that LOK "updated types" start a timer as well
a9b2de21a1528786246d353489a306b98d6f9134 changed the handling of LOK
messages to only set up an "updated" flag for types and then
when flushing the value is fetched. This means that queue() is not
called for such updates, and so the timeout to flush the messages
wasn't started.

Change-Id: I3a7324c86a55697f509f80b6c54cb949fb1daa72
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130526
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>