Advanced Diagram support: Disable enter Diagram (Groups)
Groups which are in Diagram mode should not be entered, to
do so disable at UI and - for security reasons - also at
the executing mehod.
Groups in Diagram mode can still be UnGrouped (which will
remove the Diagram status & data)

Adapted testRegenerateDiagram() to directly act on the
model data to test the same as before. Using triggering
of UI events to enter the diagram group object needs
reorganization, but model op's will always be possible

Change-Id: I199f3821c2a8308a6f015997ce9eaef87623c46a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130541
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>