Resolves: tdf#147448 ScRefreshTimerControl mutex must be std::recursive_mutex
ScRefreshTimer::Invoke() locks the mutex and subsequent nested
locks are attempted in ScDocShell::ConvertFrom() and
ScDocShellModificator ctor by ScRefreshTimerProtector.

A std::mutex must not be owned by the calling thread when a lock
is attempted, otherwise even deadlocks may occur, as was the case
here. This is exactly the difference of std::recursive_mutex.

Likely a regression from

    commit aa4aef22ab3adb74b6c9178885c4b07fee6c10eb
    CommitDate: Mon Aug 2 12:17:07 2021 +0200

        osl::Mutex->std::mutex in ScRefreshTimer

Change-Id: Iaa0f1da2b4b9616e9627d8d0001775f554756048
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130573
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins