optimize ConfigurationProperty::get()
E.g. officecfg::Office::Common::Save::ODF::DefaultVersion::get()
gets called quite often during ODS export, taking ~14% of total CPU
time, ~%5 in comphelper::getProcessComponentContext (handled by
previous commit) and the rest trying to read the property.
But it's possible to cache the final property values if there's
a notifier that clears the cache on any configuration change (it
can't be hidden as a static in the file like the other data
becase that way it'd get destroyed too late and crash). This
gets it to ~3%, most of which is spent in computing the path hash
(which could be presumably made const if the OUString hash function
were constexpr).

Change-Id: I29f3cefc8442716403d4df8bdf253f4fe847a09c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130705
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>