tdf#122471 xlsx import: pivottable error OUString, not uInt8
This fixes a LO 6.0 regression from
commit 0f3d457d2b12d01867f359c79f9408299f781b7d
which only half-way changed to OUString from uInt8.

An exception was raised because in XLSX, an INT was written
while the corresponding read function was expecting an OUString.

However, doing this runs into problems with binary files (xlsb),
which were still setting the value to an int.
Unit test shows the need to use OUString for xlsb too,
which now matches what I see in Excel 2003.

make CppunitTest_sc_pivottable_filters_test \
  CPPUNIT_TEST_NAME=testPivotTableErrorItem2FilterXLSX

Change-Id: I399c9e34984bb1ff71695a87aa56f53063d37b3d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130714
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>