Remove 'const' from for loop variable
In the commit b7c059083458d8398e2021c747ec33a731293e6e, a for loop is
converted into a range-based for loop. The pointer is used to change
the nodes using xmlUnlinkNode() and xmlFreeNode(), thus the 'const'
should have not been used.

The reinterpret_cast does not change the constness, thus I have
removed the const from the loop variable, and it will no longer be a
const reference.

Change-Id: If4c61017ea77f464230bb5802f6fc928acc7a7e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130792
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>