tdf#140077 sw_redlinehide: fix crash on SplitNode()
The problem is that CutImpl() was called with nLen=0 and so returned
without sending any hint to SwTextFrame.

Then the MergedPara gets recreated, but the problem is that the
SwParaPortion is never invalidated so it still contains the portions of
the previous merged paragraph, and painting it will crash.

But it turns out that there isn't a hint that would actually work here;
instead, generalise a condition in SwTextNode::SplitContentNode()
to simply throw away the portions.

(regression from commit 99c75736b000268331ac386c156635ad64626756)

Change-Id: I06091ca695ea9180b32e61ddb88ce64fbf20443e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131039
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>