sw clearing breaks: fix rendering of the line break char itself
The larger height is needed for the break portion, so the next line
"jumps down", below the largest fly frame. But this has the side effect,
that the line break character now has an unexpected vertical position,
as it's vertically centered.

Fix the problem by going back to the text height while painting.

No test for this, SwBreakPortion::Paint() is a NOP unless rendering on a
window, so the metafile-based rendering used for testing won't detect
the difference.

Change-Id: Ifd780c79488b149ff41d8e61585ee99c96a40128
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131093
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins