revert/fix the incorrect fix for tdf#126904
The default cell style has protection turned off. That means that
any cells that do not have a style that turns the protection on
are actually not protected. So the original code was correct,
it went beyond the cells explicitly saved in the file and the first
"default" cell was not protected, so it was ok to jump there.
Re-saving this document to xlsx and opening it in Excel has the same
result.

So revert the change. Keep the test, but edit the test file to
actually protect cells in the default cell style too, as that way
it's actually useful as a test of protected cells. And since
the test for tdf#126904 is a subset of basically the same test
for the same "bug" tdf#131455, keep only the larger test.

Change-Id: I36f21c33aaf5a31e95aa29d616c6abdec17fcf42
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131126
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>