a11y: Leave cell vector handling to AccessibleGridControlTable
Use `AccessibleGridControlTable::getAccessibleCellAt` and
cast to `AccessibleGridControlTableCell*` instead of
directly accessing the cell vector in
`AccessibleGridControl::commitCellEvent`.

`AccessibleGridControlTable::getAccessibleCellAt`
just needs row and column index as parameters,
and already takes care of everything else that's
needed.
This includes creating an accessible object for
the given indices on demand.

Therefore, limiting this to only already existing
a11y objects, which was done to avoid crashes in

    commit 128dbb4e6cf73bcd391784587bc1fcfaaf93f563
    Date:   Thu Oct 3 23:16:34 2013 +0200

        fix STL assert in accessibility::AccessibleGridControl::commitTableEvent

is no longer needed.

With this change in place, details of how cells are
organized in the vector only need to be known inside
of the `AccessibleGridControlTable` class itself, so
drop the now unused method
`AccessibleGridControlTable::getCellVector`.

(This code path is e.g. used when using
the macro from tdf#147742.)

Change-Id: I21027f0edc2904475ad6cc5fb136316f387499dc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131248
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>