(related: tdf#139687) sw: ignore following footnotes in SwTextFrameBreak
With the fix it now happens that after SwUndoDelete there is a gap at
the bottom of page 21 and the paragraph at the bottom of it is split
with 2 lines ("(here, wisdom), " etc.) on page 22.

This is because when this SwTextFrame 1927 is formatted, there are next
frames also on the page and those have footnotes on the page.

These footnotes take up space and so SwTextFrameBreak::IsInside()
breaks off the last 2 lines (1 may still fit but it has 2 widorp).

A similar problem was fixed in commit b612cfcd7828e5292259ab17377fd735851fd3b3
for footnotes anchored in follow of the current frame, but here
RemoveFootnotes() cannot be used as it requires the next frame to
already have moved to a different footnote boss, which clearly hasn't
happened yet.

So try something similar to commit 96073d87307f9f5493c3c9f2a9e543ed5e1d62f7
and count the space taken by such footnotes as available for the text
frame.

Change-Id: I8f4c1332dc471a68539306a8788a0f59d0b12b1d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131256
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>