tdf#135220 sw: fix layout after SwUndoDelete
The bugdoc is 398 pages after initial load and 397 pages after Undo.

There is really only one problem, on page 6 the fly frame 7370 "Figure
3: Hay Rake in Weeds" anchored on SwTextFrame 5896 should be on page 7
and aligned to the top of its paragraph.

First, 5986 moves forward from page 6 to 7, hitting the condition added
in commit 45517e988b508fdd0ba2076ee0c1344a6290459c
debug:1170918:1170918: SwObjectFormatterTextFrame::CheckMovedFwdCondition(): o_rbPageHasFlysAnchoredBelowThis because next chain on same page

Then 5986 moves back to page 6.  When it is formatted there, the
ConsiderObjWrapInfluenceOnObjPos() is true, preventing a formatting of
the anchor frame - so it does not move forward.

The flag SetTmpConsiderWrapInfluence(true) was set the first time (when
it moved forward).

Setting this flag for any anchor frame that moves forward was added in
2005 in commit 46297c0923c50a12510c7b15b725c8d2fd80f9c0 and apparently
the only way the flag is cleared is when SwDoc's SetModified() is
called (or the frame is in a table or column).

So it is similar to the InsertMovedFwdFrame() case that was addressed by
commit 45517e988b508fdd0ba2076ee0c1344a6290459c in that the flag should
not be set if bPageHasFlysAnchoredBelowThis is true.

(reportedly regression from sw_redlinehide
 commit 4990de6d71163ff7bd756b79a76ae347ffa1a828)

Change-Id: I27b2a1becf5f7a760e74e801e23472116d0a5ced
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131560
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>