tdf#134065, tdf#147998: fix thumbnail generation code
This moves the code drawing thumbnails and overlay icons away from
RecentDocsView to RecentDocsViewItem, where the final size of the
thumbnail is defined, so that no scaling of pre-painted icon would
happen.

This also restores the SFX_FILE_THUMBNAIL_* to state before commit
60cc9ebc078345e89779c9be10892a6d2f6cf0dc, and introduces a new set
of SFX_FILE_OVERLAY_* to be used for the overlay icons. This fixes
the display when RecentDocsThumbnail is disabled.

Change-Id: I5eb4aed5325459ddec7e196300cf6c2f9ddf80b7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131610
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>