ScriptForge - (SF_Form) fix use of BaseForm
Context: Base documents.

In a form or formcontrol event, a user script might
need the name of the form document in which the form
or control is located, f.i. to close it.

The name of the form document is supposed to be found
in the SF_Form.BaseForm property.

The property is set correctly when the Base form document
to form/formcontrol path is taken (top-down), not when the
opposite path is taken (bottom-up), typically in events.
In the latter case BaseForm is equal to an empty string.

The actual commit initializes correctly the BaseForm property
with the hierarchical name of the form document.

Change-Id: Icdee6ee9909f577f6c01ba96864fe3d6eaf3e750
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131618
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>