Resolves: tdf#148054 Advance offset for all columns, tdf#104927 regression
Regression from

    commit d2f3d45a55b695f64eb43c6475498bf70f3820ac
    CommitDate: Sat Jan 9 05:13:55 2021 +0100

        tdf#104927 consider character width for CSV import

that for fixed width introduced a visual count of character widths
for CJK IVS characters but did not advance the character offset
for skipped hidden columns.

Instead of having to ask bDetermineRange three times in the loop,
split off an extra loop for bDetermineRange, advancing the offset
then is straight forward and just only non-skipped columns are
extracted and written to document.

Change-Id: Ib6ab6b68c5fc0f6ff854d629d75742c39ed6ddf5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131722
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins