sw content controls: add document model
This is meant to represent inline structured document tags (<w:sdt>
elements in DOCX) or content controls (as Word users know this).

Don't confuse this with block-level, cell-level or table-level content
controls, which are not covered here.

You may wonder why fields or fieldmarks can't be used to represent this.
The problems are:

- a fieldmark can contain a paragraph break, inline content controls
  can't

- content controls must be a well-formed XML element, while fieldmarks
  can start/end at random document model positions

- fieldmarks are supposed to have a field command and a result result
  (with a separator between the two), but content controls don't have a
  field command

- mapping to a field has the problem that Writer fields can only have a
  single text portion / run, but content controls can have multiple ones

So model content controls with a meta-like text attribute which covers
all these cases.

SwContentControl is mostly empty as a start, it can track more data in
follow-up commits. SwTextContentControl and SwFormatContentControl are
the matching text attr and pool item subclasses.

Change-Id: I1e57f7756cf87041975371483ec4fc8abf875dfe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132291
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins